refactor(servers): consolidate client and server components for maps and plugins
Move component logic from client-side files to server components to simplify the architecture and improve performance by reducing client-side JavaScript. The changes include:
- Deleted separate client components for maps and plugins
- Integrated the logic directly into the server components
- Maintained the same functionality while reducing code duplication
