+++ app/[locale]/(main)/servers/[id]/maps/page.tsx
@@ -1,7 +1,14 @@
 import { Metadata } from 'next';
+import dynamic from 'next/dynamic';
 import { Suspense } from 'react';
+import React from 'react';
 
-import ServerMaps from '@/app/[locale]/(main)/servers/[id]/maps/page.client';
+import MapList from '@/app/[locale]/(main)/servers/[id]/maps/map-list';
+
+import ScrollContainer from '@/components/common/scroll-container';
+import Tran from '@/components/common/tran';
+import Divider from '@/components/ui/divider';
+import { ServerTabs, ServerTabsContent, ServerTabsList, ServerTabsTrigger } from '@/components/ui/server-tabs';
 
 import { Locale } from '@/i18n/config';
 import { getTranslation } from '@/i18n/server';
@@ -26,11 +33,35 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
 	};
 }
 
+const DownloadMapList = dynamic(() => import('@/app/[locale]/(main)/servers/[id]/maps/download-map-list'));
+
 export default async function Page({ params }: Props) {
 	const { id } = await params;
 	return (
-		<Suspense>
-			<ServerMaps id={id} />
-		</Suspense>
+		<div className="h-full w-full overflow-hidden p-2 flex flex-col gap-2">
+			<Suspense>
+				<ServerTabs className="gap-2" name="tab" value="list" values={['list', 'download']}>
+					<ServerTabsList className="border-none rounded-lg w-fit">
+						<ServerTabsTrigger value="list">
+							<Tran text="list" />
+						</ServerTabsTrigger>
+						<ServerTabsTrigger value="download">
+							<Tran text="download" />
+						</ServerTabsTrigger>
+					</ServerTabsList>
+					<Divider />
+					<ScrollContainer>
+						<ServerTabsContent value="list">
+							<MapList id={id} />
+						</ServerTabsContent>
+						<Suspense>
+							<ServerTabsContent value="download">
+								<DownloadMapList />
+							</ServerTabsContent>
+						</Suspense>
+					</ScrollContainer>
+				</ServerTabs>
+			</Suspense>
+		</div>
 	);
 }
